split config tests into azure test file
